# docs/openapi-spec.yaml
...
openapi: 3.0.3
info:
  title: API Système de Réservation de Salles
  description: Documentation OpenAPI générée automatiquement
  version: 1.0.0
  contact:
    name: Équipe de développement
    email: dev@company.com
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.booking.company.com
    description: Serveur de production
...
paths:
...
  /api/rooms/:
    get:
      tags:
        - rooms
      summary: Liste des salles disponibles
      description: Liste des salles disponibles - Gestion des salles
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ressource non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - rooms
      summary: Créer une nouvelle salle
      description: Créer une nouvelle salle - Gestion des salles
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ressource non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/rooms/{id}:
    get:
      tags:
        - rooms
      summary: Détails d'une salle
      description: Détails d'une salle - Gestion des salles
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ressource non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
    put:
      tags:
        - rooms
      summary: Modifier une salle
      description: Modifier une salle - Gestion des salles
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ressource non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
    delete:
      tags:
        - rooms
      summary: Supprimer une salle
      description: Supprimer une salle - Gestion des salles
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ressource non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...